+++ src/main/kotlin/kr/flooding/backend/domain/attendance/usecase/AbsenceClubMyselfUsecase.kt
@@ -3,7 +3,7 @@ package kr.flooding.backend.domain.attendance.usecase
 import kr.flooding.backend.domain.attendance.dto.request.AbsenceClubMyselfRequest
 import kr.flooding.backend.domain.attendance.persistence.entity.Attendance
 import kr.flooding.backend.domain.attendance.persistence.jpa.AttendanceJpaRepository
-import kr.flooding.backend.domain.club.persistence.entity.ClubStatus
+import kr.flooding.backend.domain.club.enums.ClubStatus
 import kr.flooding.backend.domain.club.persistence.repository.ClubRepository
 import kr.flooding.backend.domain.period.persistence.repository.PeriodRepository
 import kr.flooding.backend.global.exception.ExceptionEnum
