+++ src/main/kotlin/kr/flooding/backend/domain/classroom/controller/ClassroomController.kt
@@ -4,7 +4,7 @@ import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import kr.flooding.backend.domain.classroom.dto.request.FetchClassroomRequest
 import kr.flooding.backend.domain.classroom.dto.response.FetchClassroomResponse
-import kr.flooding.backend.domain.classroom.persistence.entity.BuildingType
+import kr.flooding.backend.domain.classroom.enums.BuildingType
 import kr.flooding.backend.domain.classroom.usecase.FetchClassroomUsecase
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
@@ -21,9 +21,9 @@ class ClassroomController(
 	@Operation(summary = "교실 조회")
 	@GetMapping
 	fun fetchClassroom(
-		@RequestParam floor: Int,
-		@RequestParam buildingType: BuildingType,
-		@RequestParam(defaultValue = "") search: String,
+        @RequestParam floor: Int,
+        @RequestParam buildingType: BuildingType,
+        @RequestParam(defaultValue = "") search: String,
 	): ResponseEntity<FetchClassroomResponse> =
 		fetchClassroomUsecase
 			.execute(
