+++ src/main/kotlin/kr/flooding/backend/domain/classroom/dto/request/FetchClassroomRequest.kt
@@ -1,9 +1,9 @@
 package kr.flooding.backend.domain.classroom.dto.request
 
-import kr.flooding.backend.domain.classroom.persistence.entity.BuildingType
+import kr.flooding.backend.domain.classroom.enums.BuildingType
 
 class FetchClassroomRequest(
-	val floor: Int,
-	val buildingType: BuildingType,
-	val search: String,
+    val floor: Int,
+    val buildingType: BuildingType,
+    val search: String,
 )
