+++ src/main/kotlin/kr/flooding/backend/domain/club/dto/common/response/ClubMyselfResponse.kt
@@ -1,19 +1,19 @@
-package kr.flooding.backend.domain.club.dto.response
+package kr.flooding.backend.domain.club.dto.common.response
 
 import kr.flooding.backend.domain.club.persistence.entity.Club
-import kr.flooding.backend.domain.club.persistence.entity.ClubStatus
-import kr.flooding.backend.domain.club.persistence.entity.ClubType
+import kr.flooding.backend.domain.club.enums.ClubStatus
+import kr.flooding.backend.domain.club.enums.ClubType
 import kr.flooding.backend.domain.user.persistence.entity.User
 import java.util.UUID
 
 class ClubMyselfResponse(
-	val id: UUID?,
-	val name: String,
-	val thumbnailImageUrl: String?,
-	val isLeader: Boolean,
-	val isRecruiting: Boolean,
-	val status: ClubStatus,
-	val type: ClubType,
+    val id: UUID?,
+    val name: String,
+    val thumbnailImageUrl: String?,
+    val isLeader: Boolean,
+    val isRecruiting: Boolean,
+    val status: ClubStatus,
+    val type: ClubType,
 ) {
 	companion object {
 		fun toDto(
