+++ src/main/kotlin/kr/flooding/backend/domain/club/dto/common/response/ClubStudentResponse.kt
@@ -1,19 +1,22 @@
-package kr.flooding.backend.domain.club.dto.response
+package kr.flooding.backend.domain.club.dto.common.response
 
 import kr.flooding.backend.domain.user.persistence.entity.StudentInfo
 import kr.flooding.backend.domain.user.persistence.entity.User
 import java.util.UUID
 
-class ClubTeacherResponse(
+class ClubStudentResponse(
 	val id: UUID?,
 	val name: String,
+	val studentInfo: StudentInfo,
 	val profileImageUrl: String?,
 ) {
 	companion object {
-		fun toDto(user: User, profileImageUrl: String?): ClubTeacherResponse {
-			return ClubTeacherResponse(
+		fun toDto(user: User, profileImageUrl: String?): ClubStudentResponse {
+			val studentInfo = requireNotNull(user.studentInfo)
+			return ClubStudentResponse(
 				id = user.id,
 				name = user.name,
+				studentInfo = studentInfo,
 				profileImageUrl = profileImageUrl,
 			)
 		}
