--- src/main/kotlin/kr/flooding/backend/domain/club/dto/response/FetchClubFilterResponse.kt
@@ -1,5 +1,0 @@
-package kr.flooding.backend.domain.club.dto.response
-
-class FetchClubFilterResponse(
-	val clubs: List<ClubFilterResponse>,
-)
