--- src/main/kotlin/kr/flooding/backend/domain/club/dto/response/FetchClubMyselfResponse.kt
@@ -1,5 +1,0 @@
-package kr.flooding.backend.domain.club.dto.response
-
-class FetchClubMyselfResponse(
-	val clubs: List<ClubMyselfResponse>,
-)
