+++ src/main/kotlin/kr/flooding/backend/domain/club/dto/web/request/UpdateClubRequest.kt
@@ -1,4 +1,4 @@
-package kr.flooding.backend.domain.club.dto.request
+package kr.flooding.backend.domain.club.dto.web.request
 
 import java.util.UUID
