+++ src/main/kotlin/kr/flooding/backend/domain/club/dto/web/response/FetchClubFilterResponse.kt
@@ -1,0 +1,7 @@
+package kr.flooding.backend.domain.club.dto.web.response
+
+import kr.flooding.backend.domain.club.dto.common.response.ClubFilterResponse
+
+class FetchClubFilterResponse(
+    val clubs: List<ClubFilterResponse>,
+)
