+++ src/main/kotlin/kr/flooding/backend/domain/club/dto/web/response/FetchClubMyselfResponse.kt
@@ -1,0 +1,7 @@
+package kr.flooding.backend.domain.club.dto.web.response
+
+import kr.flooding.backend.domain.club.dto.common.response.ClubMyselfResponse
+
+class FetchClubMyselfResponse(
+    val clubs: List<ClubMyselfResponse>,
+)
