+++ src/main/kotlin/kr/flooding/backend/domain/club/enums/ClubStatus.kt
@@ -1,0 +1,7 @@
+package kr.flooding.backend.domain.club.enums
+
+enum class ClubStatus {
+	PENDING,
+	APPROVED,
+	DENIED,
+}
