+++ src/main/kotlin/kr/flooding/backend/domain/club/enums/ClubType.kt
@@ -1,0 +1,7 @@
+package kr.flooding.backend.domain.club.enums
+
+enum class ClubType {
+	AUTONOMOUS,
+	MAJOR,
+	CAREER,
+}
