--- src/main/kotlin/kr/flooding/backend/domain/club/persistence/entity/ClubStatus.kt
@@ -1,7 +1,0 @@
-package kr.flooding.backend.domain.club.persistence.entity
-
-enum class ClubStatus {
-	PENDING,
-	APPROVED,
-	DENIED,
-}
