+++ src/main/kotlin/kr/flooding/backend/domain/club/usecase/FetchClubFilterUsecase.kt
@@ -1,9 +1,9 @@
 package kr.flooding.backend.domain.club.usecase
 
-import kr.flooding.backend.domain.club.dto.response.ClubFilterResponse
-import kr.flooding.backend.domain.club.dto.response.FetchClubFilterResponse
-import kr.flooding.backend.domain.club.persistence.entity.ClubStatus
-import kr.flooding.backend.domain.club.persistence.entity.ClubType
+import kr.flooding.backend.domain.club.dto.common.response.ClubFilterResponse
+import kr.flooding.backend.domain.club.dto.web.response.FetchClubFilterResponse
+import kr.flooding.backend.domain.club.enums.ClubStatus
+import kr.flooding.backend.domain.club.enums.ClubType
 import kr.flooding.backend.domain.club.persistence.repository.ClubRepository
 import kr.flooding.backend.global.util.S3Util
 import kr.flooding.backend.global.util.UserUtil
