+++ src/main/kotlin/kr/flooding/backend/domain/club/usecase/FetchClubMyselfUsecase.kt
@@ -1,7 +1,7 @@
 package kr.flooding.backend.domain.club.usecase
 
-import kr.flooding.backend.domain.club.dto.response.ClubMyselfResponse
-import kr.flooding.backend.domain.club.dto.response.FetchClubMyselfResponse
+import kr.flooding.backend.domain.club.dto.common.response.ClubMyselfResponse
+import kr.flooding.backend.domain.club.dto.web.response.FetchClubMyselfResponse
 import kr.flooding.backend.domain.clubMember.persistence.repository.jdsl.ClubMemberJdslRepository
 import kr.flooding.backend.global.util.S3Util
 import kr.flooding.backend.global.util.UserUtil
