+++ src/main/kotlin/kr/flooding/backend/domain/club/usecase/InviteClubMemberUsecase.kt
@@ -2,7 +2,7 @@ package kr.flooding.backend.domain.club.usecase
 
 import jakarta.transaction.Transactional
 import kr.flooding.backend.domain.club.persistence.entity.ClubInvite
-import kr.flooding.backend.domain.club.persistence.entity.ClubStatus
+import kr.flooding.backend.domain.club.enums.ClubStatus
 import kr.flooding.backend.domain.club.persistence.repository.ClubInviteRepository
 import kr.flooding.backend.domain.club.persistence.repository.ClubRepository
 import kr.flooding.backend.domain.clubMember.persistence.repository.jpa.ClubMemberJpaRepository
