+++ src/main/kotlin/kr/flooding/backend/domain/clubApplicant/persistence/repository/jpa/ClubApplicantJpaRepository.kt
@@ -1,7 +1,7 @@
 package kr.flooding.backend.domain.clubApplicant.persistence.repository.jpa
 
 import kr.flooding.backend.domain.club.persistence.entity.Club
-import kr.flooding.backend.domain.club.persistence.entity.ClubType
+import kr.flooding.backend.domain.club.enums.ClubType
 import kr.flooding.backend.domain.clubApplicant.persistence.entity.ClubApplicant
 import kr.flooding.backend.domain.user.persistence.entity.User
 import org.springframework.data.jpa.repository.JpaRepository
